Fixed bug in buildTree; DEBUG var added
Bug in buildTree() fixed: the issue with the function was related to both the assignment of its final return value and the board parameter. 

The final value returned was incorrectly stored in the score field of self.root, rather than self.root itself, which caused the entire tree to appear to not be generated (it is quite odd, however, that python did not complain about the type mismatch of the score field from int to list)

An additional mistake was passing the actual board instead of having each child manage its own unique board. While this did not cause any problem for the self.root board, all of the children boards were incorrectly resetted, causing the children to be virtually useless since the related move was not retrievable. 

Lastly, a DEBUG variable was added to facilitate the use of methods such as print_tree()