Minimax v. 1.1 Release In this version of MiniMax, critical bugs were fixed that allowed the algorithm to lose and never win.
Approach: we decided it was best to completely rewrite the class from scratch in order to pinpoint the issues in our old approach.

Fixed bugs
Bug #1: winning logic incorrect. A major issue of MiniMax v. 1 regarded the base cases of the buildTree function. This logic bug was due to an incorrect interpretation of the board state. After a winning move, the turn variable changes to the opponent that lost, but in the old basecase the turn variable was treated as the winner. This caused scores to be swapped, causing major problems in the MiniMax execution, The bug was fixed by inverting the logic of the base case. If, for example, a winning move is played and now it is X's turn, it means that O won and therefore the score should reflect the previous turn.

Bug #2: incorrect forced blocking logic. In an attempt to fix MiniMax v. 1, we tried to hard-code a simple blocking logic that would override the move MiniMax chose. At times, this resulted in MiniMax not playing optimally, which makes the tree generated unreliable. We fixed this by completely removing this feature, as it is no longer needed.

Decisions:
For now, we decided to not implement the depth feature that allowed MiniMax to pick wisely between moves with the same score. While not related to any accuracy issues, calculating the depth discriminant made the algorithm very slow. A possible fix and implementation could be adding a depth field to the TreeNode class, but this doesn't really enhance the performance of the algorithm.

Sample runs (over 100 games) results and analysis:
1. MiniMax v. Minimax: 100% ties; as expected
2. MiniMax v. GabeSimple AI: 100% ties; a bit surprising, but we noticed that MiniMax always ties with algorithm that have a blocking logic, as long as it doesn't lose it shouldn't be an issue
3. MiniMax v. RandomAI: 87% wins for MiniMax, 13% ties; as expected, MiniMax performs really well with RandomAI
4. MiniMax v. SimpleAI: 100% ties; same concept as GabeSimpleAI, maybe MiniMax is too aggressive?